requirements.txt
PyGithub==1.55
requests==2.31.0
APScheduler==3.9.1
